requirements.txt
boto3==1.34.57
fastapi==0.110.0
uvicorn==0.27.1
Jinja2==3.1.4 
requests==2.31.0
python-dotenv==1.0.1
botocore==1.34.57 
ipython==8.22.2 
pydantic==2.6.3
google-cloud-storage==2.15.0
azure-storage-blob==12.19.1
azure-keyvault-secrets==4.8.0 
azure-identity==1.15.0
azure-mgmt-compute==30.6.0  
google-compute-engine==2.8.13
google-cloud-dns==0.35.0
google-cloud-appengine-admin==1.11.3
mysql-connector-python==8.3.0
SQLAlchemy==2.0.29
pymongo==4.6.2